Version 3.28.65 (based on bleeding_edge revision r23013)
Performance and stability improvements on all platforms.

git-svn-id: https://v8.googlecode.com/svn/trunk@23014 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ build/landmines.py
@@ -58,26 +58,31 @@ def set_up_landmines(target, new_landmines):
   if not os.path.exists(out_dir):
     return
 
-  # Make sure the landmines tracker exists.
-  open(landmines_path, 'a').close()
+  if not os.path.exists(landmines_path):
+    print "Landmines tracker didn't exists."
 
-  triggered = os.path.join(out_dir, '.landmines_triggered')
-  with open(landmines_path, 'r') as f:
-    old_landmines = f.readlines()
-  if old_landmines != new_landmines:
-    old_date = time.ctime(os.stat(landmines_path).st_ctime)
-    diff = difflib.unified_diff(old_landmines, new_landmines,
-        fromfile='old_landmines', tofile='new_landmines',
-        fromfiledate=old_date, tofiledate=time.ctime(), n=0)
+  # FIXME(machenbach): Clobber deletes the .landmines tracker. Difficult
+  # to know if we are right after a clobber or if it is first-time landmines
+  # deployment. Also, a landmine-triggered clobber right after a clobber is
+  # not possible. Different clobber methods for msvs, xcode and make all
+  # have different blacklists of files that are not deleted.
+  if os.path.exists(landmines_path):
+    triggered = os.path.join(out_dir, '.landmines_triggered')
+    with open(landmines_path, 'r') as f:
+      old_landmines = f.readlines()
+    if old_landmines != new_landmines:
+      old_date = time.ctime(os.stat(landmines_path).st_ctime)
+      diff = difflib.unified_diff(old_landmines, new_landmines,
+          fromfile='old_landmines', tofile='new_landmines',
+          fromfiledate=old_date, tofiledate=time.ctime(), n=0)
 
-    with open(triggered, 'w') as f:
-      f.writelines(diff)
-    print "Setting landmine: %s" % triggered
-    print "Reason:\n%s" % diff
-  elif os.path.exists(triggered):
-    # Remove false triggered landmines.
-    os.remove(triggered)
-    print "Removing landmine: %s" % triggered
+      with open(triggered, 'w') as f:
+        f.writelines(diff)
+      print "Setting landmine: %s" % triggered
+    elif os.path.exists(triggered):
+      # Remove false triggered landmines.
+      os.remove(triggered)
+      print "Removing landmine: %s" % triggered
   with open(landmines_path, 'w') as f:
     f.writelines(new_landmines)
